# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Synapse — Kubernetes 多叢集管理平台 REST API",
        "title": "Synapse API",
        "contact": {
            "name": "Synapse Team"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/clusters": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clusters"
                ],
                "summary": "取得叢集列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "頁碼",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每頁筆數（最大 200）",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "叢集名稱搜尋",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.PagedListResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
        "/clusters/import": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clusters"
                ],
                "summary": "匯入叢集（平台管理員）",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Cluster"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
        "/clusters/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clusters"
                ],
                "summary": "取得叢集統計摘要（總叢集數、節點數、Pod 數等）",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ClusterStats"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
        "/clusters/test-connection": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clusters"
                ],
                "summary": "測試叢集 API Server 連線",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
        "/clusters/{clusterID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clusters"
                ],
                "summary": "取得叢集詳情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "叢集 ID",
                        "name": "clusterID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Cluster"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clusters"
                ],
                "summary": "刪除叢集（平台管理員）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "叢集 ID",
                        "name": "clusterID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Cluster": {
            "type": "object",
            "properties": {
                "alertmanager_config": {
                    "description": "Alertmanager 配置",
                    "type": "string"
                },
                "api_server": {
                    "type": "string"
                },
                "cert_expire_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "creator": {
                    "description": "關聯關係",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "id": {
                    "type": "integer"
                },
                "labels": {
                    "description": "JSON 格式儲存標籤",
                    "type": "string"
                },
                "last_heartbeat": {
                    "type": "string"
                },
                "monitoring_config": {
                    "description": "監控配置",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "description": "healthy, unhealthy, unknown",
                    "type": "string"
                },
                "terminal_sessions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TerminalSession"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "models.ClusterStats": {
            "type": "object",
            "properties": {
                "healthy_clusters": {
                    "type": "integer"
                },
                "ready_nodes": {
                    "type": "integer"
                },
                "running_pods": {
                    "type": "integer"
                },
                "total_clusters": {
                    "type": "integer"
                },
                "total_nodes": {
                    "type": "integer"
                },
                "total_pods": {
                    "type": "integer"
                },
                "unhealthy_clusters": {
                    "type": "integer"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "auth_type": {
                    "description": "local, ldap",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_login_at": {
                    "type": "string"
                },
                "last_login_ip": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "status": {
                    "description": "active, inactive, locked",
                    "type": "string"
                },
                "system_role": {
                    "description": "user, platform_admin, system_admin",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "response.ErrorBody": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/response.ErrorDetail"
                }
            }
        },
        "response.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.PagedListResult": {
            "type": "object",
            "properties": {
                "items": {},
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "輸入 \"Bearer {token}\" 進行認證",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}